requirements.txt
alembic==1.3.1
astroid==2.3.3
Click==7.0
ecdsa==0.14.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
gunicorn==20.0.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.4
pyasn1==0.4.8
pycryptodome==3.9.4
pylint==2.4.4
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.0.1
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.11
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2